backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.26.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-jose==3.3.0
pydantic[email]==2.7.1
pydantic-settings==2.1.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
ruff==0.1.11
black==23.12.1
